Fix how we generate style URLs.
Use styleURLs for Angular 2 components. This requires the URL to be root-relative, but the URL can't have a leading "/".

For all styles outside of src, (i.e., semantic and node_modules), we want to extract the CSS and lump it into a single
app.css, which we include in index.html.