Add support for string enums.
This is based on code in the gitbook TypeScript Deep Dive:

    https://basarat.gitbooks.io/typescript/content/docs/types/literal-types.html

The idea is to convert an array of strings into an object with those strings as the keys and values. In effect,

    const StrEnum = stringEnum(['FOO']);
    StrEnum.FOO === 'FOO'

The caller can also create a type for the enum:

    type StrEnum = keyof typeof StrEnum;

I don't see any good way to do this automatically as of TS 2.2, so it's up to the caller to create this if they want it.