Allow PascalCase (ClassCase) variable names.
The line is blurry between constructor and class in TypeScript (less than in JavaScript, though). There are cases where
we want a variable to hold a reference to a constructor, or perhaps to a string enum object. In these cases, we want the
object to have the same name as the type, which means the variable should use PascalCase.